Remove grunt-concurrent
Was mostly just confusing and bugprone to run the devserver and the tasks
in the same tasks, especially if the devserver crashed and you needed to
restart the watch task as well to get it going again.

New solution, just run the devserver in one terminal, and watch in
another.